requirements.txt
aiohttp
async-timeout
beautifulsoup4
bs4
certifi
chardet
discord.py
google-api-python-client
httplib2
idna
multidict
oauth2client
pyasn1
pyasn1-modules
python-dateutil
regex
rsa
six
uritemplate
urllib3
websockets
wikipedia
youtube-dl
pillow
